Added role based program update/delete, fixed directory front, fixed role based front view

+++ bug-bounty-api/src/program/program.service.ts
@@ -1,15 +1,19 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { UserEntity } from 'src/auth/entities/user.entity';
 import { UserRoleEnum } from 'src/enums/user-role-enum';
 import { Repository } from 'typeorm';
 import { AddProgramDto } from './dto/addProgram.dto';
 import { ProgramEntity } from './entities/program.entity';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class ProgramService {
     constructor(
         @InjectRepository(ProgramEntity)
-        private programRepository: Repository<ProgramEntity>
+        private programRepository: Repository<ProgramEntity>,
+        @InjectRepository(UserEntity)
+        private userRepository: Repository<UserEntity>,
     ){}
     async getPrograms(company): Promise<ProgramEntity[]>{
         if(company.role === UserRoleEnum.ADMIN){
@@ -26,29 +30,51 @@ export class ProgramService {
     async getOneProgram(searchedId : number): Promise<ProgramEntity>{
         return await this.programRepository.findOne({id:searchedId});
     }
-
-    async updateProgram(id:string,program: AddProgramDto):Promise<ProgramEntity>{
-
+    async getDirectory(): Promise<Partial<ProgramEntity[]>>{
+        return await this.programRepository.find();
+    }
+    async updateProgram(id:string,program: AddProgramDto,company):Promise<ProgramEntity>{
+        
         const updateProgram = await this.programRepository.findOne(+id);
-        if (program.name.length > 0) {
-            updateProgram.name = program.name
+        if(!updateProgram){
+            throw new NotFoundException("Program not found");
         }
-        if (program.policy.length > 0){
-            updateProgram.policy = program.policy
+        console.log(updateProgram.company.id);
+        console.log(company.id);
+        if(updateProgram.company.id == company.id){
+            if (program.name.length > 0) {
+                updateProgram.name = program.name
+            }
+            if (program.policy.length > 0){
+                updateProgram.policy = program.policy
+            }
+            updateProgram.updatedAt = new Date()
+            if (program.inScope.length > 0) {
+                updateProgram.inScope=program.inScope;
+            }
+            if (program.outOfScope.length > 0) {
+                updateProgram.outOfScope=program.outOfScope;
+            }
+            return await this.programRepository.save(updateProgram);
         }
-        updateProgram.updatedAt = new Date()
-        if (program.inScope.length > 0) {
-            updateProgram.inScope=program.inScope;
+        else  {
+            throw new UnauthorizedException("Unauthorized")
         }
-        if (program.outOfScope.length > 0) {
-            updateProgram.outOfScope=program.outOfScope;
-        }
-        return await this.programRepository.save(updateProgram);
     }
 
 
-    async deleteProgram(id):Promise<ProgramEntity>{
-        const deletedProgram = await this.programRepository.findOne(+id);
-        return await this.programRepository.remove(deletedProgram);
+    async deleteProgram(password,id,user):Promise<Partial<ProgramEntity>>{
+        const u = await this.userRepository.findOne(user.id);
+        
+        const hashedPassword = await bcrypt.hash(password,u.salt);
+        if(hashedPassword === u.password){
+            const deletedProgram = await this.programRepository.findOne(+id);
+            return await this.programRepository.remove(deletedProgram);
+        }
+        else {
+            throw new UnauthorizedException('Incorrect password');
+        }
+        
+        
     }
 }
